refactor(openclaw): extract shared RPC helpers to rpc.ts
Move RpcRequest, RpcResponse, isRpcResponse, buildRpcRequest,
openWebSocket, safeCloseWebSocket, and rpcCall into a shared
module to eliminate duplication between execute.ts and test.ts.
